Refactor Abstract*VersionStore* test classes (PR 2/2)
Move tests out of `AbstractDatabaseAdapterVersionStoreTest` and `AbstractVersionStoreTestBase`
into base classes containing groups of tests (e.g. related to commit-op, related to commit-log-op,
etc).
