+++ versioned/tests/src/main/java/org/projectnessie/versioned/tests/AbstractVersionStoreTestBase.java
@@ -15,1746 +15,75 @@
  */
 package org.projectnessie.versioned.tests;
 
-import static java.util.Collections.singletonList;
-import static org.assertj.core.api.Assertions.assertThat;
-import static org.assertj.core.api.Assertions.assertThatThrownBy;
-import static org.junit.jupiter.api.Assertions.assertEquals;
-import static org.junit.jupiter.api.Assertions.assertNotEquals;
-import static org.junit.jupiter.api.Assertions.assertThrows;
-import static org.junit.jupiter.api.Assertions.assertTrue;
-
-import com.google.common.collect.ImmutableList;
-import com.google.common.collect.ImmutableMap;
-import com.google.common.collect.Lists;
-import java.util.ArrayList;
-import java.util.Arrays;
-import java.util.Collections;
-import java.util.List;
-import java.util.Optional;
-import java.util.function.Function;
-import java.util.function.IntFunction;
-import java.util.stream.Collectors;
-import java.util.stream.IntStream;
-import java.util.stream.Stream;
-import org.junit.jupiter.api.BeforeEach;
-import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Nested;
-import org.junit.jupiter.api.Test;
-import org.junit.jupiter.params.ParameterizedTest;
-import org.junit.jupiter.params.provider.MethodSource;
-import org.projectnessie.versioned.BranchName;
-import org.projectnessie.versioned.Commit;
-import org.projectnessie.versioned.Delete;
-import org.projectnessie.versioned.Diff;
-import org.projectnessie.versioned.GetNamedRefsParams;
-import org.projectnessie.versioned.Hash;
-import org.projectnessie.versioned.Key;
-import org.projectnessie.versioned.Put;
-import org.projectnessie.versioned.Ref;
-import org.projectnessie.versioned.ReferenceAlreadyExistsException;
-import org.projectnessie.versioned.ReferenceConflictException;
-import org.projectnessie.versioned.ReferenceInfo;
-import org.projectnessie.versioned.ReferenceNotFoundException;
 import org.projectnessie.versioned.StringStoreWorker;
-import org.projectnessie.versioned.TagName;
-import org.projectnessie.versioned.Unchanged;
 import org.projectnessie.versioned.VersionStore;
-import org.projectnessie.versioned.VersionStoreException;
-import org.projectnessie.versioned.WithHash;
-import org.projectnessie.versioned.WithType;
 
 /** Base class used for integration tests against version store implementations. */
 public abstract class AbstractVersionStoreTestBase {
 
   protected abstract VersionStore<String, String, StringStoreWorker.TestEnum> store();
 
-  /** Use case simulation: single branch, multiple users, each user updating a separate table. */
-  @Test
-  protected void singleBranchManyUsersDistinctTables() throws Exception {
-    singleBranchTest(
-        "singleBranchManyUsersDistinctTables", user -> String.format("user-table-%d", user), false);
-  }
-
-  /** Use case simulation: single branch, multiple users, all users updating a single table. */
-  @Test
-  protected void singleBranchManyUsersSingleTable() throws Exception {
-    singleBranchTest("singleBranchManyUsersSingleTable", user -> "single-table", true);
-  }
-
-  private void singleBranchTest(
-      String branchName, IntFunction<String> tableNameGen, boolean allowInconsistentValueException)
-      throws Exception {
-    BranchName branch = BranchName.of(branchName);
-
-    int numUsers = 5;
-    int numCommits = 50;
-
-    Hash[] hashesKnownByUser = new Hash[numUsers];
-    Hash createHash = store().create(branch, Optional.empty());
-    Arrays.fill(hashesKnownByUser, createHash);
-
-    List<String> expectedValues = new ArrayList<>();
-    for (int commitNum = 0; commitNum < numCommits; commitNum++) {
-      for (int user = 0; user < numUsers; user++) {
-        Hash hashKnownByUser = hashesKnownByUser[user];
-
-        String msg = String.format("user %03d/commit %03d", user, commitNum);
-        expectedValues.add(msg);
-        String value = String.format("data_file_%03d_%03d", user, commitNum);
-        Put<String> put = Put.of(Key.of(tableNameGen.apply(user)), value);
-
-        Hash commitHash;
-        try {
-          commitHash =
-              store().commit(branch, Optional.of(hashKnownByUser), msg, ImmutableList.of(put));
-        } catch (ReferenceConflictException inconsistentValueException) {
-          if (allowInconsistentValueException) {
-            hashKnownByUser = store().hashOnReference(branch, Optional.empty());
-            commitHash =
-                store().commit(branch, Optional.of(hashKnownByUser), msg, ImmutableList.of(put));
-          } else {
-            throw inconsistentValueException;
-          }
-        }
-
-        assertNotEquals(hashKnownByUser, commitHash);
-
-        hashesKnownByUser[user] = commitHash;
-      }
-    }
-
-    // Verify that all commits are there and that the order of the commits is correct
-    List<String> committedValues = commitsList(branch, s -> s.map(Commit::getCommitMeta), false);
-    Collections.reverse(expectedValues);
-    assertEquals(expectedValues, committedValues);
-  }
-
-  /*
-   * Test:
-   * - Create a branch with no hash assigned to it
-   * - check that a hash is returned by toHash
-   * - check the branch is returned by getNamedRefs
-   * - check that no commits are returned using getCommits
-   * - check the branch cannot be created
-   * - check the branch can be deleted
-   */
-  @Test
-  public void createAndDeleteBranch() throws Exception {
-    final BranchName branch = BranchName.of("foo");
-    store().create(branch, Optional.empty());
-    final Hash hash = store().hashOnReference(branch, Optional.empty());
-    assertThat(hash).isNotNull();
-
-    final BranchName anotherBranch = BranchName.of("bar");
-    final Hash createHash = store().create(anotherBranch, Optional.of(hash));
-    final Hash commitHash = commit("Some Commit").toBranch(anotherBranch);
-    assertNotEquals(createHash, commitHash);
-
-    final BranchName anotherAnotherBranch = BranchName.of("baz");
-    final Hash otherCreateHash = store().create(anotherAnotherBranch, Optional.of(commitHash));
-    assertEquals(commitHash, otherCreateHash);
-
-    List<ReferenceInfo<String>> namedRefs;
-    try (Stream<ReferenceInfo<String>> str =
-        store().getNamedRefs(GetNamedRefsParams.DEFAULT).filter(this::filterMainBranch)) {
-      namedRefs = str.collect(Collectors.toList());
-    }
-    assertThat(namedRefs)
-        .containsExactlyInAnyOrder(
-            ReferenceInfo.of(hash, branch),
-            ReferenceInfo.of(commitHash, anotherBranch),
-            ReferenceInfo.of(commitHash, anotherAnotherBranch));
-
-    assertThat(commitsList(branch, false)).isEmpty();
-    assertThat(commitsList(anotherBranch, false)).hasSize(1);
-    assertThat(commitsList(anotherAnotherBranch, false)).hasSize(1);
-    assertThat(commitsList(hash, false)).isEmpty(); // empty commit should not be listed
-    assertThat(commitsList(commitHash, false)).hasSize(1); // empty commit should not be listed
-
-    assertThrows(
-        ReferenceAlreadyExistsException.class, () -> store().create(branch, Optional.empty()));
-    assertThrows(
-        ReferenceAlreadyExistsException.class, () -> store().create(branch, Optional.of(hash)));
-
-    store().delete(branch, Optional.of(hash));
-    assertThrows(
-        ReferenceNotFoundException.class, () -> store().hashOnReference(branch, Optional.empty()));
-    try (Stream<ReferenceInfo<String>> str =
-        store().getNamedRefs(GetNamedRefsParams.DEFAULT).filter(this::filterMainBranch)) {
-      assertThat(str).hasSize(2); // bar + baz
-    }
-    assertThrows(ReferenceNotFoundException.class, () -> store().delete(branch, Optional.of(hash)));
-  }
-
-  /** Exclude {@code main} branch in tests. */
-  private boolean filterMainBranch(ReferenceInfo<String> r) {
-    return !r.getNamedRef().getName().equals("main");
-  }
-
-  @Test
-  public void commitLogPaging() throws Exception {
-    BranchName branch = BranchName.of("commitLogPaging");
-    Hash createHash = store().create(branch, Optional.empty());
-
-    int commits = 95; // this should be enough
-    Hash[] commitHashes = new Hash[commits];
-    List<String> messages = new ArrayList<>(commits);
-    for (int i = 0; i < commits; i++) {
-      String msg = String.format("commit#%05d", i);
-      messages.add(msg);
-      commitHashes[i] =
-          store()
-              .commit(
-                  branch,
-                  Optional.of(i == 0 ? createHash : commitHashes[i - 1]),
-                  msg,
-                  ImmutableList.of(Put.of(Key.of("table"), String.format("value#%05d", i))));
-    }
-    Collections.reverse(messages);
-
-    List<String> justTwo = commitsList(branch, s -> s.limit(2).map(Commit::getCommitMeta), false);
-    assertEquals(messages.subList(0, 2), justTwo);
-    List<String> justTen = commitsList(branch, s -> s.limit(10).map(Commit::getCommitMeta), false);
-    assertEquals(messages.subList(0, 10), justTen);
-
-    int pageSize = 10;
-
-    // Test parameter sanity check. Want the last page to be smaller than the page-size.
-    assertNotEquals(0, commits % (pageSize - 1));
-
-    Hash lastHash = null;
-    for (int offset = 0; ; ) {
-      List<Commit<String, String>> logPage =
-          commitsList(lastHash == null ? branch : lastHash, s -> s.limit(pageSize), false);
-
-      assertEquals(
-          messages.subList(offset, Math.min(offset + pageSize, commits)),
-          logPage.stream().map(Commit::getCommitMeta).collect(Collectors.toList()));
-
-      lastHash = logPage.get(logPage.size() - 1).getHash();
-
-      offset += pageSize - 1;
-      if (offset >= commits) {
-        // The "next after last page" should always return just a single commit, that's basically
-        // the "end of commit-log"-condition.
-        logPage = commitsList(lastHash, s -> s.limit(pageSize), false);
-        assertEquals(
-            Collections.singletonList(messages.get(commits - 1)),
-            logPage.stream().map(Commit::getCommitMeta).collect(Collectors.toList()));
-        break;
-      }
-    }
-  }
-
-  /*
-   * Test:
-   * - Create a branch with no hash assigned to it
-   * - add a commit to the branch
-   * - create a tag for the initial hash
-   * - create another tag for the hash after the commit
-   * - check that cannot create existing tags, or tag with no assigned hash
-   * - check that a hash is returned by toHash
-   * - check the tags are returned by getNamedRefs
-   * - check that expected commits are returned by getCommits
-   * - check the branch can be deleted
-   */
-  @Test
-  public void createAndDeleteTag() throws Exception {
-    final BranchName branch = BranchName.of("foo");
-    store().create(branch, Optional.empty());
-
-    final Hash initialHash = store().hashOnReference(branch, Optional.empty());
-    final Hash commitHash = commit("Some commit").toBranch(branch);
-
-    final TagName tag = TagName.of("tag");
-    store().create(tag, Optional.of(initialHash));
-
-    final TagName anotherTag = TagName.of("another-tag");
-    store().create(anotherTag, Optional.of(commitHash));
-
-    assertThrows(
-        ReferenceAlreadyExistsException.class, () -> store().create(tag, Optional.of(initialHash)));
-
-    assertThat(store().hashOnReference(tag, Optional.empty())).isEqualTo(initialHash);
-    assertThat(store().hashOnReference(anotherTag, Optional.empty())).isEqualTo(commitHash);
-
-    List<ReferenceInfo<String>> namedRefs;
-    try (Stream<ReferenceInfo<String>> str =
-        store().getNamedRefs(GetNamedRefsParams.DEFAULT).filter(this::filterMainBranch)) {
-      namedRefs = str.collect(Collectors.toList());
-    }
-    assertThat(namedRefs)
-        .containsExactlyInAnyOrder(
-            ReferenceInfo.of(commitHash, branch),
-            ReferenceInfo.of(initialHash, tag),
-            ReferenceInfo.of(commitHash, anotherTag));
-
-    assertThat(commitsList(tag, false)).isEmpty();
-    assertThat(commitsList(initialHash, false)).isEmpty(); // empty commit should not be listed
-
-    assertThat(commitsList(anotherTag, false)).hasSize(1);
-    assertThat(commitsList(commitHash, false)).hasSize(1); // empty commit should not be listed
-
-    store().delete(tag, Optional.of(initialHash));
-    assertThrows(
-        ReferenceNotFoundException.class, () -> store().hashOnReference(tag, Optional.empty()));
-    try (Stream<ReferenceInfo<String>> str =
-        store().getNamedRefs(GetNamedRefsParams.DEFAULT).filter(this::filterMainBranch)) {
-      assertThat(str).hasSize(2); // foo + another-tag
-    }
-    assertThrows(
-        ReferenceNotFoundException.class, () -> store().delete(tag, Optional.of(initialHash)));
-  }
-
-  /*
-   * Test:
-   * - Create a new branch
-   * - Add a commit to it
-   * - Check that another commit with no operations can be added with the initial hash
-   * - Check the commit can be listed
-   * - Check that the commit can be deleted
-   */
-  @Test
-  public void commitToBranch() throws Exception {
-    final BranchName branch = BranchName.of("foo");
-
-    final Hash createHash = store().create(branch, Optional.empty());
-    final Hash initialHash = store().hashOnReference(branch, Optional.empty());
-    assertEquals(createHash, initialHash);
-
-    final Hash commitHash0 =
-        store().commit(branch, Optional.of(initialHash), "Some commit", Collections.emptyList());
-    final Hash commitHash = store().hashOnReference(branch, Optional.empty());
-    assertEquals(commitHash, commitHash0);
-
-    assertThat(commitHash).isNotEqualTo(initialHash);
-    store().commit(branch, Optional.of(initialHash), "Another commit", Collections.emptyList());
-    final Hash anotherCommitHash = store().hashOnReference(branch, Optional.empty());
-
-    assertThat(commitsList(branch, false))
-        .contains(commit(anotherCommitHash, "Another commit"), commit(commitHash, "Some commit"));
-    assertThat(commitsList(commitHash, false)).contains(commit(commitHash, "Some commit"));
-
-    assertThrows(
-        ReferenceConflictException.class, () -> store().delete(branch, Optional.of(initialHash)));
-    store().delete(branch, Optional.of(anotherCommitHash));
-    assertThrows(
-        ReferenceNotFoundException.class, () -> store().hashOnReference(branch, Optional.empty()));
-    try (Stream<ReferenceInfo<String>> str =
-        store().getNamedRefs(GetNamedRefsParams.DEFAULT).filter(this::filterMainBranch)) {
-      assertThat(str).isEmpty();
-    }
-    assertThrows(
-        ReferenceNotFoundException.class, () -> store().delete(branch, Optional.of(commitHash)));
-  }
-
-  /*
-   * Test:
-   * - Create a new branch
-   * - Add 3 commits in succession with no conflicts to it with put and delete operations
-   * - Check commit metadata
-   * - Check keys for each commit hash
-   * - Check values for each commit hash
-   */
-  @Test
-  public void commitSomeOperations() throws Exception {
-    final BranchName branch = BranchName.of("foo");
-
-    store().create(branch, Optional.empty());
-
-    final Hash initialCommit =
-        commit("Initial Commit")
-            .put("t1", "v1_1")
-            .put("t2", "v2_1")
-            .put("t3", "v3_1")
-            .toBranch(branch);
-
-    final Hash secondCommit =
-        commit("Second Commit")
-            .put("t1", "v1_2")
-            .delete("t2")
-            .delete("t3")
-            .put("t4", "v4_1")
-            .toBranch(branch);
-
-    final Hash thirdCommit =
-        commit("Third Commit").put("t2", "v2_2").unchanged("t4").toBranch(branch);
-
-    assertThat(commitsList(branch, false))
-        .contains(
-            commit(thirdCommit, "Third Commit"),
-            commit(secondCommit, "Second Commit"),
-            commit(initialCommit, "Initial Commit"));
-
-    try (Stream<Key> keys = store().getKeys(branch).map(WithType::getValue)) {
-      assertThat(keys).containsExactlyInAnyOrder(Key.of("t1"), Key.of("t2"), Key.of("t4"));
-    }
-
-    try (Stream<Key> keys = store().getKeys(secondCommit).map(WithType::getValue)) {
-      assertThat(keys).containsExactlyInAnyOrder(Key.of("t1"), Key.of("t4"));
-    }
-
-    try (Stream<Key> keys = store().getKeys(initialCommit).map(WithType::getValue)) {
-      assertThat(keys).containsExactlyInAnyOrder(Key.of("t1"), Key.of("t2"), Key.of("t3"));
-    }
-
-    assertThat(
-            store()
-                .getValues(
-                    secondCommit,
-                    Arrays.asList(Key.of("t1"), Key.of("t2"), Key.of("t3"), Key.of("t4"))))
-        .containsExactlyInAnyOrderEntriesOf(
-            ImmutableMap.of(Key.of("t1"), "v1_2", Key.of("t4"), "v4_1"));
-
-    assertThat(
-            store()
-                .getValues(
-                    initialCommit,
-                    Arrays.asList(Key.of("t1"), Key.of("t2"), Key.of("t3"), Key.of("t4"))))
-        .containsExactlyInAnyOrderEntriesOf(
-            ImmutableMap.of(
-                Key.of("t1"), "v1_1",
-                Key.of("t2"), "v2_1",
-                Key.of("t3"), "v3_1"));
-
-    assertThat(store().getValue(branch, Key.of("t1"))).isEqualTo("v1_2");
-    assertThat(store().getValue(branch, Key.of("t2"))).isEqualTo("v2_2");
-    assertThat(store().getValue(branch, Key.of("t3"))).isNull();
-    assertThat(store().getValue(branch, Key.of("t4"))).isEqualTo("v4_1");
-
-    assertThat(store().getValue(secondCommit, Key.of("t1"))).isEqualTo("v1_2");
-    assertThat(store().getValue(secondCommit, Key.of("t2"))).isNull();
-    assertThat(store().getValue(secondCommit, Key.of("t3"))).isNull();
-    assertThat(store().getValue(secondCommit, Key.of("t4"))).isEqualTo("v4_1");
-
-    assertThat(store().getValue(initialCommit, Key.of("t1"))).isEqualTo("v1_1");
-    assertThat(store().getValue(initialCommit, Key.of("t2"))).isEqualTo("v2_1");
-    assertThat(store().getValue(initialCommit, Key.of("t3"))).isEqualTo("v3_1");
-    assertThat(store().getValue(initialCommit, Key.of("t4"))).isNull();
-  }
-
-  /*
-   * Test:
-   * - Create a new branch
-   * - Add a commit for 3 keys
-   * - Add a commit based on initial commit for first key
-   * - Add a commit based on initial commit for second key
-   * - Add a commit based on initial commit for third  key
-   * - Check commit metadata
-   * - Check keys for each commit hash
-   * - Check values for each commit hash
-   */
-  @Test
-  public void commitNonConflictingOperations() throws Exception {
-    final BranchName branch = BranchName.of("foo");
-
-    store().create(branch, Optional.empty());
-
-    final Hash initialCommit =
-        commit("Initial Commit")
-            .put("t1", "v1_1")
-            .put("t2", "v2_1")
-            .put("t3", "v3_1")
-            .toBranch(branch);
-
-    final Hash t1Commit =
-        commit("T1 Commit").fromReference(initialCommit).put("t1", "v1_2").toBranch(branch);
-    final Hash t2Commit =
-        commit("T2 Commit").fromReference(initialCommit).delete("t2").toBranch(branch);
-    final Hash t3Commit =
-        commit("T3 Commit").fromReference(initialCommit).unchanged("t3").toBranch(branch);
-    final Hash extraCommit =
-        commit("Extra Commit")
-            .fromReference(t1Commit)
-            .put("t1", "v1_3")
-            .put("t3", "v3_2")
-            .toBranch(branch);
-    final Hash newT2Commit =
-        commit("New T2 Commit").fromReference(t2Commit).put("t2", "new_v2_1").toBranch(branch);
-
-    assertThat(commitsList(branch, false))
-        .contains(
-            commit(newT2Commit, "New T2 Commit"),
-            commit(extraCommit, "Extra Commit"),
-            commit(t3Commit, "T3 Commit"),
-            commit(t2Commit, "T2 Commit"),
-            commit(t1Commit, "T1 Commit"),
-            commit(initialCommit, "Initial Commit"));
-
-    try (Stream<Key> keys = store().getKeys(branch).map(WithType::getValue)) {
-      assertThat(keys).containsExactlyInAnyOrder(Key.of("t1"), Key.of("t2"), Key.of("t3"));
-    }
-
-    assertThat(store().getValues(branch, Arrays.asList(Key.of("t1"), Key.of("t2"), Key.of("t3"))))
-        .containsExactlyInAnyOrderEntriesOf(
-            ImmutableMap.of(
-                Key.of("t1"), "v1_3",
-                Key.of("t2"), "new_v2_1",
-                Key.of("t3"), "v3_2"));
-
-    assertThat(
-            store().getValues(newT2Commit, Arrays.asList(Key.of("t1"), Key.of("t2"), Key.of("t3"))))
-        .containsExactlyInAnyOrderEntriesOf(
-            ImmutableMap.of(
-                Key.of("t1"), "v1_3",
-                Key.of("t2"), "new_v2_1",
-                Key.of("t3"), "v3_2"));
-
-    assertThat(
-            store().getValues(extraCommit, Arrays.asList(Key.of("t1"), Key.of("t2"), Key.of("t3"))))
-        .containsExactlyInAnyOrderEntriesOf(
-            ImmutableMap.of(
-                Key.of("t1"), "v1_3",
-                Key.of("t3"), "v3_2"));
-
-    assertThat(store().getValues(t3Commit, Arrays.asList(Key.of("t1"), Key.of("t2"), Key.of("t3"))))
-        .containsExactlyInAnyOrderEntriesOf(
-            ImmutableMap.of(
-                Key.of("t1"), "v1_2",
-                Key.of("t3"), "v3_1"));
-
-    assertThat(store().getValues(t2Commit, Arrays.asList(Key.of("t1"), Key.of("t2"), Key.of("t3"))))
-        .containsExactlyInAnyOrderEntriesOf(
-            ImmutableMap.of(
-                Key.of("t1"), "v1_2",
-                Key.of("t3"), "v3_1"));
-
-    assertThat(store().getValues(t1Commit, Arrays.asList(Key.of("t1"), Key.of("t2"), Key.of("t3"))))
-        .containsExactlyInAnyOrderEntriesOf(
-            ImmutableMap.of(
-                Key.of("t1"), "v1_2",
-                Key.of("t2"), "v2_1",
-                Key.of("t3"), "v3_1"));
-  }
-
-  /*
-   * Test:
-   * - Create a new branch
-   * - Add a commit to create 2 keys
-   * - Add a second commit to delete one key and add a new one
-   * - Check that put operations against 1st commit for the 3 keys fail
-   * - Check that delete operations against 1st commit for the 3 keys fail
-   * - Check that unchanged operations against 1st commit for the 3 keys fail
-   * - Check that branch state hasn't changed
-   */
-  @Test
-  public void commitConflictingOperations() throws Exception {
-    final BranchName branch = BranchName.of("foo");
-
-    store().create(branch, Optional.empty());
-
-    final Hash initialCommit =
-        commit("Initial Commit").put("t1", "v1_1").put("t2", "v2_1").toBranch(branch);
-
-    final Hash secondCommit =
-        commit("Second Commit").put("t1", "v1_2").delete("t2").put("t3", "v3_1").toBranch(branch);
-
-    assertThrows(
-        ReferenceConflictException.class,
-        () ->
-            commit("Conflicting Commit")
-                .fromReference(initialCommit)
-                .put("t1", "v1_3")
-                .toBranch(branch));
-    assertThrows(
-        ReferenceConflictException.class,
-        () ->
-            commit("Conflicting Commit")
-                .fromReference(initialCommit)
-                .put("t2", "v2_2")
-                .toBranch(branch));
-    assertThrows(
-        ReferenceConflictException.class,
-        () ->
-            commit("Conflicting Commit")
-                .fromReference(initialCommit)
-                .put("t3", "v3_2")
-                .toBranch(branch));
-
-    assertThrows(
-        ReferenceConflictException.class,
-        () ->
-            commit("Conflicting Commit")
-                .fromReference(initialCommit)
-                .delete("t1")
-                .toBranch(branch));
-    assertThrows(
-        ReferenceConflictException.class,
-        () ->
-            commit("Conflicting Commit")
-                .fromReference(initialCommit)
-                .delete("t2")
-                .toBranch(branch));
-    assertThrows(
-        ReferenceConflictException.class,
-        () ->
-            commit("Conflicting Commit")
-                .fromReference(initialCommit)
-                .delete("t3")
-                .toBranch(branch));
-
-    // Checking the state hasn't changed
-    assertThat(store().hashOnReference(branch, Optional.empty())).isEqualTo(secondCommit);
-  }
-
-  /*
-   * Test:
-   * - Create a new branch
-   * - Add a commit to create 2 keys
-   * - Add a second commit to delete one key and add a new one
-   * - force commit put operations
-   * - Check that put operations against 1st commit for the 3 keys fail
-   * - Check that delete operations against 1st commit for the 3 keys fail
-   * - Check that unchanged operations against 1st commit for the 3 keys fail
-   * - Check that branch state hasn't changed
-   */
-  @Test
-  public void forceCommitConflictingOperations() throws Exception {
-    final BranchName branch = BranchName.of("foo");
-
-    store().create(branch, Optional.empty());
-
-    commit("Initial Commit").put("t1", "v1_1").put("t2", "v2_1").toBranch(branch);
-
-    commit("Second Commit").put("t1", "v1_2").delete("t2").put("t3", "v3_1").toBranch(branch);
-
-    final Hash putCommit =
-        forceCommit("Conflicting Commit")
-            .put("t1", "v1_3")
-            .put("t2", "v2_2")
-            .put("t3", "v3_2")
-            .toBranch(branch);
-
-    assertThat(store().hashOnReference(branch, Optional.empty())).isEqualTo(putCommit);
-    assertThat(store().getValues(branch, Arrays.asList(Key.of("t1"), Key.of("t2"), Key.of("t3"))))
-        .containsExactlyInAnyOrderEntriesOf(
-            ImmutableMap.of(
-                Key.of("t1"), "v1_3",
-                Key.of("t2"), "v2_2",
-                Key.of("t3"), "v3_2"));
-
-    final Hash unchangedCommit =
-        commit("Conflicting Commit")
-            .unchanged("t1")
-            .unchanged("t2")
-            .unchanged("t3")
-            .toBranch(branch);
-    assertThat(store().hashOnReference(branch, Optional.empty())).isEqualTo(unchangedCommit);
-    assertThat(store().getValues(branch, Arrays.asList(Key.of("t1"), Key.of("t2"), Key.of("t3"))))
-        .containsExactlyInAnyOrderEntriesOf(
-            ImmutableMap.of(
-                Key.of("t1"), "v1_3",
-                Key.of("t2"), "v2_2",
-                Key.of("t3"), "v3_2"));
-
-    final Hash deleteCommit =
-        commit("Conflicting Commit").delete("t1").delete("t2").delete("t3").toBranch(branch);
-    assertThat(store().hashOnReference(branch, Optional.empty())).isEqualTo(deleteCommit);
-    assertThat(store().getValues(branch, Arrays.asList(Key.of("t1"), Key.of("t2"), Key.of("t3"))))
-        .isEmpty();
-  }
-
-  /*
-   * Test:
-   *  - Check that store allows storing the same value under different keys
-   */
-  @Test
-  public void commitDuplicateValues() throws Exception {
-    BranchName branch = BranchName.of("dupe-values");
-    store().create(branch, Optional.empty());
-    store()
-        .commit(
-            branch,
-            Optional.empty(),
-            "metadata",
-            ImmutableList.of(put("keyA", "foo"), put("keyB", "foo")));
-
-    assertThat(store().getValue(branch, Key.of("keyA"))).isEqualTo("foo");
-    assertThat(store().getValue(branch, Key.of("keyB"))).isEqualTo("foo");
-  }
-
-  /*
-   * Test:
-   * - Check that store throws RNFE if branch doesn't exist
-   */
-  @Test
-  public void commitWithInvalidBranch() {
-    final BranchName branch = BranchName.of("unknown");
-
-    assertThrows(
-        ReferenceNotFoundException.class,
-        () -> store().commit(branch, Optional.empty(), "New commit", Collections.emptyList()));
-  }
-
-  /*
-   * Test:
-   * - Check that store throws RNFE if reference hash doesn't exist
-   */
-  @Test
-  public void commitWithUnknownReference()
-      throws ReferenceNotFoundException, ReferenceAlreadyExistsException {
-    final BranchName branch = BranchName.of("foo");
-    store().create(branch, Optional.empty());
-
-    assertThrows(
-        ReferenceNotFoundException.class,
-        () ->
-            store()
-                .commit(
-                    branch,
-                    Optional.of(Hash.of("1234567890abcdef")),
-                    "New commit",
-                    Collections.emptyList()));
-  }
-
-  /*
-   * Test:
-   * - Check that store throws IllegalArgumentException if reference hash is not in branch ancestry
-   */
-  @Test
-  public void commitWithInvalidReference()
-      throws ReferenceNotFoundException, ReferenceConflictException,
-          ReferenceAlreadyExistsException {
-    final BranchName branch = BranchName.of("foo");
-    store().create(branch, Optional.empty());
-
-    final Hash initialHash = store().hashOnReference(branch, Optional.empty());
-    store().commit(branch, Optional.of(initialHash), "Some commit", Collections.emptyList());
-
-    final Hash commitHash = store().hashOnReference(branch, Optional.empty());
-
-    final BranchName branch2 = BranchName.of("bar");
-    store().create(branch2, Optional.empty());
-
-    assertThrows(
-        ReferenceNotFoundException.class,
-        () ->
-            store()
-                .commit(
-                    branch2, Optional.of(commitHash), "Another commit", Collections.emptyList()));
-  }
-
-  @Test
-  public void getValueForEmptyBranch()
-      throws ReferenceNotFoundException, ReferenceAlreadyExistsException {
-    BranchName branch = BranchName.of("empty-branch");
-    store().create(branch, Optional.empty());
-    final Hash hash = store().hashOnReference(branch, Optional.empty());
-
-    assertThat(store().getValue(hash, Key.of("arbitrary"))).isNull();
-  }
-
-  @Test
-  public void assign() throws VersionStoreException {
-    final BranchName branch = BranchName.of("foo");
-    store().create(branch, Optional.empty());
-    final Hash initialHash = store().hashOnReference(branch, Optional.empty());
-
-    final Hash commit = commit("Some commit").toBranch(branch);
-    store().create(BranchName.of("bar"), Optional.of(commit));
-    store().create(TagName.of("tag1"), Optional.of(commit));
-    store().create(TagName.of("tag2"), Optional.of(commit));
-    store().create(TagName.of("tag3"), Optional.of(commit));
-
-    final Hash anotherCommit = commit("Another commit").toBranch(branch);
-    store().assign(TagName.of("tag2"), Optional.of(commit), anotherCommit);
-    store().assign(TagName.of("tag3"), Optional.empty(), anotherCommit);
-
-    assertThrows(
-        ReferenceNotFoundException.class,
-        () -> store().assign(BranchName.of("baz"), Optional.empty(), anotherCommit));
-    assertThrows(
-        ReferenceNotFoundException.class,
-        () -> store().assign(TagName.of("unknowon-tag"), Optional.empty(), anotherCommit));
-
-    assertThrows(
-        ReferenceConflictException.class,
-        () -> store().assign(TagName.of("tag1"), Optional.of(initialHash), commit));
-    assertThrows(
-        ReferenceConflictException.class,
-        () -> store().assign(TagName.of("tag1"), Optional.of(initialHash), anotherCommit));
-    assertThrows(
-        ReferenceNotFoundException.class,
-        () -> store().assign(TagName.of("tag1"), Optional.of(commit), Hash.of("1234567890abcdef")));
-
-    assertThat(commitsList(branch, false))
-        .contains(commit(anotherCommit, "Another commit"), commit(commit, "Some commit"));
-
-    assertThat(commitsList(BranchName.of("bar"), false)).contains(commit(commit, "Some commit"));
-
-    assertThat(commitsList(TagName.of("tag1"), false)).contains(commit(commit, "Some commit"));
-
-    assertThat(commitsList(TagName.of("tag2"), false))
-        .contains(commit(anotherCommit, "Another commit"), commit(commit, "Some commit"));
-  }
-
-  static void assertCommitMeta(
-      List<Commit<String, String>> current,
-      List<Commit<String, String>> expected,
-      Function<String, String> commitMetaModifier) {
-    assertThat(current)
-        .map(Commit::getCommitMeta)
-        .containsExactlyElementsOf(
-            expected.stream()
-                .map(Commit::getCommitMeta)
-                .map(commitMetaModifier)
-                .collect(Collectors.toList()));
-  }
-
   @Nested
-  @DisplayName("when transplanting")
-  protected class WhenTransplanting {
-
-    private Hash initialHash;
-    private Hash firstCommit;
-    private Hash secondCommit;
-    private Hash thirdCommit;
-    private List<Commit<String, String>> commits;
-
-    String transplanted(String commitMeta) {
-      return commitMeta + ", transplanted";
-    }
-
-    @BeforeEach
-    protected void setupCommits() throws VersionStoreException {
-      final BranchName branch = BranchName.of("foo");
-      store().create(branch, Optional.empty());
-
-      initialHash = store().hashOnReference(branch, Optional.empty());
-
-      firstCommit =
-          commit("Initial Commit")
-              .put("t1", "v1_1")
-              .put("t2", "v2_1")
-              .put("t3", "v3_1")
-              .toBranch(branch);
-
-      secondCommit =
-          commit("Second Commit")
-              .put("t1", "v1_2")
-              .delete("t2")
-              .delete("t3")
-              .put("t4", "v4_1")
-              .toBranch(branch);
-
-      thirdCommit = commit("Third Commit").put("t2", "v2_2").unchanged("t4").toBranch(branch);
-
-      commits = commitsList(branch, false);
-    }
-
-    @Test
-    protected void checkTransplantOnEmptyBranch() throws VersionStoreException {
-      checkTransplantOnEmptyBranch(Function.identity());
-    }
-
-    @Test
-    protected void checkTransplantOnEmptyBranchModify() throws VersionStoreException {
-      checkTransplantOnEmptyBranch(this::transplanted);
-    }
-
-    private void checkTransplantOnEmptyBranch(Function<String, String> commitMetaModify)
-        throws VersionStoreException {
-      final BranchName newBranch = BranchName.of("bar_1");
-      store().create(newBranch, Optional.empty());
-
-      store()
-          .transplant(
-              newBranch,
-              Optional.of(initialHash),
-              Arrays.asList(firstCommit, secondCommit, thirdCommit),
-              commitMetaModify);
-      assertThat(
-              store()
-                  .getValues(
-                      newBranch,
-                      Arrays.asList(Key.of("t1"), Key.of("t2"), Key.of("t3"), Key.of("t4"))))
-          .containsExactlyInAnyOrderEntriesOf(
-              ImmutableMap.of(
-                  Key.of("t1"), "v1_2",
-                  Key.of("t2"), "v2_2",
-                  Key.of("t4"), "v4_1"));
-
-      assertCommitMeta(commitsList(newBranch, false).subList(0, 3), commits, commitMetaModify);
-    }
-
-    @Test
-    protected void checkTransplantWithPreviousCommit() throws VersionStoreException {
-      final BranchName newBranch = BranchName.of("bar_2");
-      store().create(newBranch, Optional.empty());
-      commit("Unrelated commit").put("t5", "v5_1").toBranch(newBranch);
-
-      store()
-          .transplant(
-              newBranch,
-              Optional.of(initialHash),
-              Arrays.asList(firstCommit, secondCommit, thirdCommit),
-              Function.identity());
-      assertThat(
-              store()
-                  .getValues(
-                      newBranch,
-                      Arrays.asList(
-                          Key.of("t1"), Key.of("t2"), Key.of("t3"), Key.of("t4"), Key.of("t5"))))
-          .containsExactlyInAnyOrderEntriesOf(
-              ImmutableMap.of(
-                  Key.of("t1"), "v1_2",
-                  Key.of("t2"), "v2_2",
-                  Key.of("t4"), "v4_1",
-                  Key.of("t5"), "v5_1"));
-    }
-
-    @Test
-    protected void checkTransplantWitConflictingCommit() throws VersionStoreException {
-      final BranchName newBranch = BranchName.of("bar_3");
-      store().create(newBranch, Optional.empty());
-      commit("Another commit").put("t1", "v1_4").toBranch(newBranch);
-
-      assertThrows(
-          ReferenceConflictException.class,
-          () ->
-              store()
-                  .transplant(
-                      newBranch,
-                      Optional.of(initialHash),
-                      Arrays.asList(firstCommit, secondCommit, thirdCommit),
-                      Function.identity()));
-    }
-
-    @Test
-    protected void checkTransplantWithDelete() throws VersionStoreException {
-      final BranchName newBranch = BranchName.of("bar_4");
-      store().create(newBranch, Optional.empty());
-      commit("Another commit").put("t1", "v1_4").toBranch(newBranch);
-      commit("Another commit").delete("t1").toBranch(newBranch);
-
-      store()
-          .transplant(
-              newBranch,
-              Optional.of(initialHash),
-              Arrays.asList(firstCommit, secondCommit, thirdCommit),
-              Function.identity());
-      assertThat(
-              store()
-                  .getValues(
-                      newBranch,
-                      Arrays.asList(Key.of("t1"), Key.of("t2"), Key.of("t3"), Key.of("t4"))))
-          .containsExactlyInAnyOrderEntriesOf(
-              ImmutableMap.of(
-                  Key.of("t1"), "v1_2",
-                  Key.of("t2"), "v2_2",
-                  Key.of("t4"), "v4_1"));
-    }
-
-    @Test
-    protected void checkTransplantOnNonExistingBranch() {
-      final BranchName newBranch = BranchName.of("bar_5");
-      assertThrows(
-          ReferenceNotFoundException.class,
-          () ->
-              store()
-                  .transplant(
-                      newBranch,
-                      Optional.of(initialHash),
-                      Arrays.asList(firstCommit, secondCommit, thirdCommit),
-                      Function.identity()));
-    }
-
-    @Test
-    protected void checkTransplantWithNonExistingCommit() throws VersionStoreException {
-      final BranchName newBranch = BranchName.of("bar_6");
-      store().create(newBranch, Optional.empty());
-      assertThrows(
-          ReferenceNotFoundException.class,
-          () ->
-              store()
-                  .transplant(
-                      newBranch,
-                      Optional.of(initialHash),
-                      Collections.singletonList(Hash.of("1234567890abcdef")),
-                      Function.identity()));
-    }
-
-    @Test
-    protected void checkTransplantWithNoExpectedHash() throws VersionStoreException {
-      final BranchName newBranch = BranchName.of("bar_7");
-      store().create(newBranch, Optional.empty());
-      commit("Another commit").put("t5", "v5_1").toBranch(newBranch);
-      commit("Another commit").put("t1", "v1_4").toBranch(newBranch);
-
-      store()
-          .transplant(
-              newBranch,
-              Optional.empty(),
-              Arrays.asList(firstCommit, secondCommit, thirdCommit),
-              Function.identity());
-      assertThat(
-              store()
-                  .getValues(
-                      newBranch,
-                      Arrays.asList(
-                          Key.of("t1"), Key.of("t2"), Key.of("t3"), Key.of("t4"), Key.of("t5"))))
-          .containsExactlyInAnyOrderEntriesOf(
-              ImmutableMap.of(
-                  Key.of("t1"), "v1_2",
-                  Key.of("t2"), "v2_2",
-                  Key.of("t4"), "v4_1",
-                  Key.of("t5"), "v5_1"));
-    }
-
-    @Test
-    protected void checkTransplantWithCommitsInWrongOrder() throws VersionStoreException {
-      final BranchName newBranch = BranchName.of("bar_8");
-      store().create(newBranch, Optional.empty());
-
-      assertThrows(
-          IllegalArgumentException.class,
-          () ->
-              store()
-                  .transplant(
-                      newBranch,
-                      Optional.empty(),
-                      Arrays.asList(secondCommit, firstCommit, thirdCommit),
-                      Function.identity()));
-    }
-
-    @Test
-    protected void checkInvalidBranchHash() throws VersionStoreException {
-      final BranchName anotherBranch = BranchName.of("bar");
-      store().create(anotherBranch, Optional.empty());
-      final Hash unrelatedCommit =
-          commit("Another Commit")
-              .put("t1", "v1_1")
-              .put("t2", "v2_1")
-              .put("t3", "v3_1")
-              .toBranch(anotherBranch);
-
-      final BranchName newBranch = BranchName.of("bar_1");
-      store().create(newBranch, Optional.empty());
-
-      assertThrows(
-          ReferenceNotFoundException.class,
-          () ->
-              store()
-                  .transplant(
-                      newBranch,
-                      Optional.of(unrelatedCommit),
-                      Arrays.asList(firstCommit, secondCommit, thirdCommit),
-                      Function.identity()));
-    }
-
-    @Test
-    protected void transplantBasic() throws VersionStoreException {
-      final BranchName newBranch = BranchName.of("bar_2");
-      store().create(newBranch, Optional.empty());
-      commit("Unrelated commit").put("t5", "v5_1").toBranch(newBranch);
-
-      store()
-          .transplant(
-              newBranch,
-              Optional.of(initialHash),
-              Arrays.asList(firstCommit, secondCommit),
-              Function.identity());
-      assertThat(
-              store().getValues(newBranch, Arrays.asList(Key.of("t1"), Key.of("t4"), Key.of("t5"))))
-          .containsExactlyInAnyOrderEntriesOf(
-              ImmutableMap.of(
-                  Key.of("t1"), "v1_2",
-                  Key.of("t4"), "v4_1",
-                  Key.of("t5"), "v5_1"));
+  public class Commits extends AbstractCommits {
+    public Commits() {
+      super(AbstractVersionStoreTestBase.this.store());
     }
   }
 
   @Nested
-  protected class WhenMerging {
-
-    private Hash initialHash;
-    private Hash firstCommit;
-    private Hash secondCommit;
-    private Hash thirdCommit;
-    private List<Commit<String, String>> commits;
-
-    @BeforeEach
-    protected void setupCommits() throws VersionStoreException {
-      final BranchName branch = BranchName.of("foo");
-      store().create(branch, Optional.empty());
-
-      // The default common ancestor for all merge-tests.
-      // The spec for 'VersionStore.merge' mentions "(...) until we arrive at a common ancestor",
-      // but old implementations allowed a merge even if the "merge-from" and "merge-to" have no
-      // common ancestor and did merge "everything" from the "merge-from" into "merge-to".
-      // Note: "beginning-of-time" (aka creating a branch without specifying a "create-from")
-      // creates a new commit-tree that is decoupled from other commit-trees.
-      initialHash = commit("Default common ancestor").toBranch(branch);
-
-      firstCommit =
-          commit("First Commit")
-              .put("t1", "v1_1")
-              .put("t2", "v2_1")
-              .put("t3", "v3_1")
-              .toBranch(branch);
-      secondCommit =
-          commit("Second Commit")
-              .put("t1", "v1_2")
-              .delete("t2")
-              .delete("t3")
-              .put("t4", "v4_1")
-              .toBranch(branch);
-      thirdCommit = commit("Third Commit").put("t2", "v2_2").unchanged("t4").toBranch(branch);
-
-      commits = commitsList(branch, false).subList(0, 3);
-    }
-
-    private String merged(String commitMeta) {
-      return commitMeta + ", merged";
-    }
-
-    @Test
-    protected void mergeIntoEmptyBranch() throws VersionStoreException {
-      final BranchName newBranch = BranchName.of("mergeIntoEmptyBranch");
-      store().create(newBranch, Optional.of(initialHash));
-
-      store().merge(thirdCommit, newBranch, Optional.of(initialHash), Function.identity());
-      assertThat(
-              store()
-                  .getValues(
-                      newBranch,
-                      Arrays.asList(Key.of("t1"), Key.of("t2"), Key.of("t3"), Key.of("t4"))))
-          .containsExactlyInAnyOrderEntriesOf(
-              ImmutableMap.of(
-                  Key.of("t1"), "v1_2",
-                  Key.of("t2"), "v2_2",
-                  Key.of("t4"), "v4_1"));
-
-      // not modifying commit meta, will just "fast forward"
-      assertThat(store().hashOnReference(newBranch, Optional.empty())).isEqualTo(thirdCommit);
-
-      assertCommitMeta(commitsList(newBranch, false).subList(0, 3), commits, Function.identity());
-    }
-
-    @Test
-    protected void mergeIntoEmptyBranchModifying() throws VersionStoreException {
-      final BranchName newBranch = BranchName.of("mergeIntoEmptyBranchModifying");
-      store().create(newBranch, Optional.of(initialHash));
-
-      store().merge(thirdCommit, newBranch, Optional.of(initialHash), this::merged);
-      assertThat(
-              store()
-                  .getValues(
-                      newBranch,
-                      Arrays.asList(Key.of("t1"), Key.of("t2"), Key.of("t3"), Key.of("t4"))))
-          .containsExactlyInAnyOrderEntriesOf(
-              ImmutableMap.of(
-                  Key.of("t1"), "v1_2",
-                  Key.of("t2"), "v2_2",
-                  Key.of("t4"), "v4_1"));
-
-      // modify the commit meta, will generate new commits and therefore new commit hashes
-      assertThat(store().hashOnReference(newBranch, Optional.empty())).isNotEqualTo(thirdCommit);
-
-      assertCommitMeta(commitsList(newBranch, false).subList(0, 3), commits, this::merged);
-    }
-
-    @Test
-    protected void mergeIntoNonConflictingBranch() throws VersionStoreException {
-      final BranchName newBranch = BranchName.of("bar_2");
-      store().create(newBranch, Optional.of(initialHash));
-      final Hash newCommit = commit("Unrelated commit").put("t5", "v5_1").toBranch(newBranch);
-
-      store().merge(thirdCommit, newBranch, Optional.empty(), Function.identity());
-      assertThat(
-              store()
-                  .getValues(
-                      newBranch,
-                      Arrays.asList(
-                          Key.of("t1"), Key.of("t2"), Key.of("t3"), Key.of("t4"), Key.of("t5"))))
-          .containsExactlyInAnyOrderEntriesOf(
-              ImmutableMap.of(
-                  Key.of("t1"), "v1_2",
-                  Key.of("t2"), "v2_2",
-                  Key.of("t4"), "v4_1",
-                  Key.of("t5"), "v5_1"));
-
-      final List<Commit<String, String>> commits = commitsList(newBranch, false);
-      assertThat(commits)
-          .satisfiesExactly(
-              c0 -> assertThat(c0).extracting(Commit::getCommitMeta).isEqualTo("Third Commit"),
-              c1 -> assertThat(c1).extracting(Commit::getCommitMeta).isEqualTo("Second Commit"),
-              c2 -> assertThat(c2).extracting(Commit::getCommitMeta).isEqualTo("First Commit"),
-              c3 -> assertThat(c3).extracting(Commit::getHash).isEqualTo(newCommit),
-              c4 -> assertThat(c4).extracting(Commit::getHash).isEqualTo(initialHash));
-    }
-
-    @Test
-    protected void nonEmptyFastForwardMerge() throws VersionStoreException {
-      final Key key = Key.of("t1");
-      final BranchName etl = BranchName.of("etl");
-      final BranchName review = BranchName.of("review");
-      store().create(etl, Optional.of(initialHash));
-      store().create(review, Optional.of(initialHash));
-      store()
-          .commit(
-              etl, Optional.empty(), "commit 1", Collections.singletonList(Put.of(key, "value1")));
-      store()
-          .merge(
-              store().hashOnReference(etl, Optional.empty()),
-              review,
-              Optional.empty(),
-              Function.identity());
-      store()
-          .commit(
-              etl, Optional.empty(), "commit 2", Collections.singletonList(Put.of(key, "value2")));
-      store()
-          .merge(
-              store().hashOnReference(etl, Optional.empty()),
-              review,
-              Optional.empty(),
-              Function.identity());
-      assertEquals(store().getValue(review, key), "value2");
-    }
-
-    @Test
-    protected void mergeWithCommonAncestor() throws VersionStoreException {
-      final BranchName newBranch = BranchName.of("bar_2");
-      store().create(newBranch, Optional.of(firstCommit));
-
-      final Hash newCommit = commit("Unrelated commit").put("t5", "v5_1").toBranch(newBranch);
-
-      store().merge(thirdCommit, newBranch, Optional.empty(), Function.identity());
-      assertThat(
-              store()
-                  .getValues(
-                      newBranch,
-                      Arrays.asList(
-                          Key.of("t1"), Key.of("t2"), Key.of("t3"), Key.of("t4"), Key.of("t5"))))
-          .containsExactlyInAnyOrderEntriesOf(
-              ImmutableMap.of(
-                  Key.of("t1"), "v1_2",
-                  Key.of("t2"), "v2_2",
-                  Key.of("t4"), "v4_1",
-                  Key.of("t5"), "v5_1"));
-
-      final List<Commit<String, String>> commits = commitsList(newBranch, false);
-      assertThat(commits).hasSize(5);
-      assertThat(commits.get(4).getHash()).isEqualTo(initialHash);
-      assertThat(commits.get(3).getHash()).isEqualTo(firstCommit);
-      assertThat(commits.get(2).getHash()).isEqualTo(newCommit);
-      assertThat(commits.get(1).getCommitMeta()).isEqualTo("Second Commit");
-      assertThat(commits.get(0).getCommitMeta()).isEqualTo("Third Commit");
-    }
-
-    @Test
-    protected void mergeWithConflictingKeys() throws VersionStoreException {
-      final BranchName foo = BranchName.of("foofoo");
-      final BranchName bar = BranchName.of("barbar");
-      store().create(foo, Optional.of(this.initialHash));
-      store().create(bar, Optional.of(this.initialHash));
-
-      // we're essentially modifying the same key on both branches and then merging one branch into
-      // the other and expect a conflict
-      Key key1 = Key.of("some_key1");
-      Key key2 = Key.of("some_key2");
-
-      store()
-          .commit(
-              foo, Optional.empty(), "commit 1", Collections.singletonList(Put.of(key1, "value1")));
-      store()
-          .commit(
-              bar, Optional.empty(), "commit 2", Collections.singletonList(Put.of(key1, "value2")));
-      store()
-          .commit(
-              foo, Optional.empty(), "commit 3", Collections.singletonList(Put.of(key2, "value3")));
-      Hash barHash =
-          store()
-              .commit(
-                  bar,
-                  Optional.empty(),
-                  "commit 4",
-                  Collections.singletonList(Put.of(key2, "value4")));
-
-      assertThatThrownBy(() -> store().merge(barHash, foo, Optional.empty(), Function.identity()))
-          .isInstanceOf(ReferenceConflictException.class)
-          .hasMessageContaining("The following keys have been changed in conflict:")
-          .hasMessageContaining(key1.toString())
-          .hasMessageContaining(key2.toString());
-    }
-
-    @Test
-    protected void mergeIntoConflictingBranch() throws VersionStoreException {
-      final BranchName newBranch = BranchName.of("bar_3");
-      store().create(newBranch, Optional.of(initialHash));
-      commit("Another commit").put("t1", "v1_4").toBranch(newBranch);
-
-      assertThrows(
-          ReferenceConflictException.class,
-          () ->
-              store().merge(thirdCommit, newBranch, Optional.of(initialHash), Function.identity()));
-    }
-
-    @Test
-    protected void mergeIntoNonExistingBranch() {
-      final BranchName newBranch = BranchName.of("bar_5");
-      assertThrows(
-          ReferenceNotFoundException.class,
-          () ->
-              store().merge(thirdCommit, newBranch, Optional.of(initialHash), Function.identity()));
-    }
-
-    @Test
-    protected void mergeIntoNonExistingReference() throws VersionStoreException {
-      final BranchName newBranch = BranchName.of("bar_6");
-      store().create(newBranch, Optional.of(initialHash));
-      assertThrows(
-          ReferenceNotFoundException.class,
-          () ->
-              store()
-                  .merge(
-                      Hash.of("1234567890abcdef"),
-                      newBranch,
-                      Optional.of(initialHash),
-                      Function.identity()));
+  public class Contents extends AbstractContents {
+    public Contents() {
+      super(AbstractVersionStoreTestBase.this.store());
     }
   }
 
-  @Test
-  void toRef() throws VersionStoreException {
-    final BranchName branch = BranchName.of("toRef");
-    store().create(branch, Optional.empty());
-    store().hashOnReference(branch, Optional.empty());
-
-    final Hash firstCommit = commit("First Commit").toBranch(branch);
-
-    assertThat(store().toRef(branch.getName())).isEqualTo(WithHash.of(firstCommit, branch));
-
-    final Hash secondCommit = commit("Second Commit").toBranch(branch);
-    final Hash thirdCommit = commit("Third Commit").toBranch(branch);
-
-    store().create(BranchName.of(thirdCommit.asString()), Optional.of(firstCommit));
-    store().create(TagName.of(secondCommit.asString()), Optional.of(firstCommit));
-
-    assertThat(store().toRef(secondCommit.asString()))
-        .isEqualTo(WithHash.of(firstCommit, TagName.of(secondCommit.asString())));
-    assertThat(store().toRef(thirdCommit.asString()))
-        .isEqualTo(WithHash.of(firstCommit, BranchName.of(thirdCommit.asString())));
-    // Is it correct to allow a reference with the sentinel reference?
-    // assertThat(store().toRef(initialCommit.asString()), is(WithHash.of(initialCommit,
-    // initialCommit)));
-    assertThrows(ReferenceNotFoundException.class, () -> store().toRef("unknown-ref"));
-    assertThrows(ReferenceNotFoundException.class, () -> store().toRef("1234567890abcdef"));
-  }
-
-  @Test
-  protected void checkDiff() throws VersionStoreException {
-    final BranchName branch = BranchName.of("checkDiff");
-    store().create(branch, Optional.empty());
-    final Hash initial = store().hashOnReference(branch, Optional.empty());
-
-    final Hash firstCommit =
-        commit("First Commit").put("k1", "v1").put("k2", "v2").toBranch(branch);
-    final Hash secondCommit =
-        commit("Second Commit").put("k2", "v2a").put("k3", "v3").toBranch(branch);
-
-    List<Diff<String>> startToSecond =
-        store().getDiffs(initial, secondCommit).collect(Collectors.toList());
-    assertThat(startToSecond)
-        .containsExactlyInAnyOrder(
-            Diff.of(Key.of("k1"), Optional.empty(), Optional.of("v1")),
-            Diff.of(Key.of("k2"), Optional.empty(), Optional.of("v2a")),
-            Diff.of(Key.of("k3"), Optional.empty(), Optional.of("v3")));
-
-    List<Diff<String>> secondToStart =
-        store().getDiffs(secondCommit, initial).collect(Collectors.toList());
-    assertThat(secondToStart)
-        .containsExactlyInAnyOrder(
-            Diff.of(Key.of("k1"), Optional.of("v1"), Optional.empty()),
-            Diff.of(Key.of("k2"), Optional.of("v2a"), Optional.empty()),
-            Diff.of(Key.of("k3"), Optional.of("v3"), Optional.empty()));
-
-    List<Diff<String>> firstToSecond =
-        store().getDiffs(firstCommit, secondCommit).collect(Collectors.toList());
-    assertThat(firstToSecond)
-        .containsExactlyInAnyOrder(
-            Diff.of(Key.of("k2"), Optional.of("v2"), Optional.of("v2a")),
-            Diff.of(Key.of("k3"), Optional.empty(), Optional.of("v3")));
-
-    List<Diff<String>> firstToFirst =
-        store().getDiffs(firstCommit, firstCommit).collect(Collectors.toList());
-    assertTrue(firstToFirst.isEmpty());
-  }
-
-  @Test
-  public void commitLogExtendedNoGlobalState() throws Exception {
-    BranchName branch = BranchName.of("commitLogExtended");
-    Hash firstParent = store().create(branch, Optional.empty());
-
-    int numCommits = 10;
-
-    List<Hash> hashes =
-        IntStream.rangeClosed(1, numCommits)
-            .mapToObj(
-                i -> {
-                  try {
-                    return commit("Commit #" + i)
-                        .put("k" + i, StringStoreWorker.withId("v" + i, "c" + i))
-                        .put("key" + i, StringStoreWorker.withId("value" + i, "cid" + i))
-                        .delete("delete" + i)
-                        .toBranch(branch);
-                  } catch (Exception e) {
-                    throw new RuntimeException(e);
-                  }
-                })
-            .collect(Collectors.toList());
-    List<Hash> parentHashes =
-        Stream.concat(Stream.of(firstParent), hashes.subList(0, 9).stream())
-            .collect(Collectors.toList());
-
-    assertThat(Lists.reverse(commitsList(branch, false)))
-        .allSatisfy(
-            c -> {
-              assertThat(c.getOperations()).isNull();
-              assertThat(c.getParentHash()).isNull();
-            })
-        .extracting(Commit::getHash)
-        .containsExactlyElementsOf(hashes);
-
-    List<Commit<String, String>> commits = Lists.reverse(commitsList(branch, true));
-    assertThat(IntStream.rangeClosed(1, numCommits))
-        .allSatisfy(
-            i -> {
-              Commit<String, String> c = commits.get(i - 1);
-              assertThat(c)
-                  .extracting(
-                      Commit::getCommitMeta,
-                      Commit::getHash,
-                      Commit::getParentHash,
-                      Commit::getOperations)
-                  .containsExactly(
-                      "Commit #" + i,
-                      hashes.get(i - 1),
-                      parentHashes.get(i - 1),
-                      Arrays.asList(
-                          Delete.of(Key.of("delete" + i)),
-                          Put.of(
-                              Key.of("k" + i),
-                              StringStoreWorker.withStateAndId("v" + i, "v" + i, "c" + i)),
-                          Put.of(
-                              Key.of("key" + i),
-                              StringStoreWorker.withStateAndId(
-                                  "value" + i, "value" + i, "cid" + i))));
-            });
-  }
-
-  protected CommitBuilder<String, String, StringStoreWorker.TestEnum> forceCommit(String message) {
-    return new CommitBuilder<>(store()).withMetadata(message);
-  }
-
-  protected CommitBuilder<String, String, StringStoreWorker.TestEnum> commit(String message) {
-    return new CommitBuilder<>(store()).withMetadata(message).fromLatest();
-  }
-
-  protected Put<String> put(String key, String value) {
-    return Put.of(Key.of(key), value);
-  }
-
-  protected Delete<String> delete(String key) {
-    return Delete.of(Key.of(key));
-  }
-
-  protected Unchanged<String> unchanged(String key) {
-    return Unchanged.of(Key.of(key));
-  }
-
-  protected List<Commit<String, String>> commitsList(Ref ref, boolean fetchAdditionalInfo)
-      throws ReferenceNotFoundException {
-    return commitsList(ref, Function.identity(), fetchAdditionalInfo);
-  }
-
-  protected <T> List<T> commitsList(
-      Ref ref,
-      Function<Stream<Commit<String, String>>, Stream<T>> streamFunction,
-      boolean fetchAdditionalInfo)
-      throws ReferenceNotFoundException {
-    try (Stream<Commit<String, String>> s = store().getCommits(ref, fetchAdditionalInfo)) {
-      return streamFunction.apply(s).collect(Collectors.toList());
+  @Nested
+  public class CommitLog extends AbstractCommitLog {
+    public CommitLog() {
+      super(AbstractVersionStoreTestBase.this.store());
     }
   }
 
-  protected static Commit<String, String> commit(Hash hash, String commitMeta) {
-    return Commit.<String, String>builder().hash(hash).commitMeta(commitMeta).build();
-  }
-
-  /** Test behavior when a table (content-key) is created, then dropped and then re-created. */
-  @Test
-  void recreateTable() throws Exception {
-    BranchName branch = BranchName.of("recreateTable-main");
-    Key key = Key.of("recreateTable");
-
-    store().create(branch, Optional.empty());
-    // commit just something to have a "real" common ancestor and not "beginning of time", which
-    // means no-common-ancestor
-    Hash ancestor =
-        store()
-            .commit(
-                branch,
-                Optional.empty(),
-                "create table",
-                singletonList(
-                    Put.of(
-                        key,
-                        StringStoreWorker.withStateAndId(
-                            "initial-state", "value", "CONTENT-ID-1"))));
-    assertThat(store().getValue(branch, key)).isEqualTo("initial-state|value@CONTENT-ID-1");
-    assertThat(store().getValue(ancestor, key)).isEqualTo("initial-state|value@CONTENT-ID-1");
-
-    Hash delete =
-        store().commit(branch, Optional.empty(), "drop table", ImmutableList.of(Delete.of(key)));
-    assertThat(store().getValue(branch, key)).isNull();
-    assertThat(store().getValue(delete, key)).isNull();
-
-    Hash recreate =
-        store()
-            .commit(
-                branch,
-                Optional.empty(),
-                "drop table",
-                ImmutableList.of(
-                    Put.of(
-                        key,
-                        StringStoreWorker.withStateAndId(
-                            "recreate-state", "value", "CONTENT-ID-DIFFERENT"))));
-    assertThat(store().getValue(branch, key))
-        .isEqualTo("recreate-state|value@CONTENT-ID-DIFFERENT");
-    assertThat(store().getValue(recreate, key))
-        .isEqualTo("recreate-state|value@CONTENT-ID-DIFFERENT");
-  }
-
-  private static class ReferenceNotFoundFunction {
-
-    final String name;
-    String msg;
-    ThrowingFunction setup;
-    ThrowingFunction function;
-
-    ReferenceNotFoundFunction(String name) {
-      this.name = name;
-    }
-
-    ReferenceNotFoundFunction setup(ThrowingFunction setup) {
-      this.setup = setup;
-      return this;
-    }
-
-    ReferenceNotFoundFunction function(ThrowingFunction function) {
-      this.function = function;
-      return this;
-    }
-
-    ReferenceNotFoundFunction msg(String msg) {
-      this.msg = msg;
-      return this;
-    }
-
-    @Override
-    public String toString() {
-      return name;
-    }
-
-    @FunctionalInterface
-    interface ThrowingFunction {
-      void run(VersionStore<String, String, StringStoreWorker.TestEnum> store)
-          throws VersionStoreException;
+  @Nested
+  public class References extends AbstractReferences {
+    public References() {
+      super(AbstractVersionStoreTestBase.this.store());
     }
   }
 
-  static List<ReferenceNotFoundFunction> referenceNotFoundFunctions() {
-    return Arrays.asList(
-        // getCommits()
-        new ReferenceNotFoundFunction("getCommits/branch")
-            .msg("Named reference 'this-one-should-not-exist' not found")
-            .function(s -> s.getCommits(BranchName.of("this-one-should-not-exist"), false)),
-        new ReferenceNotFoundFunction("getCommits/tag")
-            .msg("Named reference 'this-one-should-not-exist' not found")
-            .function(s -> s.getCommits(TagName.of("this-one-should-not-exist"), false)),
-        new ReferenceNotFoundFunction("getCommits/hash")
-            .msg("Commit '12341234123412341234123412341234123412341234' not found")
-            .function(
-                s -> s.getCommits(Hash.of("12341234123412341234123412341234123412341234"), false)),
-        // getValue()
-        new ReferenceNotFoundFunction("getValue/branch")
-            .msg("Named reference 'this-one-should-not-exist' not found")
-            .function(s -> s.getValue(BranchName.of("this-one-should-not-exist"), Key.of("foo"))),
-        new ReferenceNotFoundFunction("getValue/tag")
-            .msg("Named reference 'this-one-should-not-exist' not found")
-            .function(s -> s.getValue(TagName.of("this-one-should-not-exist"), Key.of("foo"))),
-        new ReferenceNotFoundFunction("getValue/hash")
-            .msg("Commit '12341234123412341234123412341234123412341234' not found")
-            .function(
-                s ->
-                    s.getValue(
-                        Hash.of("12341234123412341234123412341234123412341234"), Key.of("foo"))),
-        // getValues()
-        new ReferenceNotFoundFunction("getValues/branch")
-            .msg("Named reference 'this-one-should-not-exist' not found")
-            .function(
-                s ->
-                    s.getValues(
-                        BranchName.of("this-one-should-not-exist"), singletonList(Key.of("foo")))),
-        new ReferenceNotFoundFunction("getValues/tag")
-            .msg("Named reference 'this-one-should-not-exist' not found")
-            .function(
-                s ->
-                    s.getValues(
-                        TagName.of("this-one-should-not-exist"), singletonList(Key.of("foo")))),
-        new ReferenceNotFoundFunction("getValues/hash")
-            .msg("Commit '12341234123412341234123412341234123412341234' not found")
-            .function(
-                s ->
-                    s.getValues(
-                        Hash.of("12341234123412341234123412341234123412341234"),
-                        singletonList(Key.of("foo")))),
-        // getKeys()
-        new ReferenceNotFoundFunction("getKeys/branch")
-            .msg("Named reference 'this-one-should-not-exist' not found")
-            .function(s -> s.getKeys(BranchName.of("this-one-should-not-exist"))),
-        new ReferenceNotFoundFunction("getKeys/tag")
-            .msg("Named reference 'this-one-should-not-exist' not found")
-            .function(s -> s.getKeys(TagName.of("this-one-should-not-exist"))),
-        new ReferenceNotFoundFunction("getKeys/hash")
-            .msg("Commit '12341234123412341234123412341234123412341234' not found")
-            .function(s -> s.getKeys(Hash.of("12341234123412341234123412341234123412341234"))),
-        // assign()
-        new ReferenceNotFoundFunction("assign/branch/ok")
-            .msg("Named reference 'this-one-should-not-exist' not found")
-            .function(
-                s ->
-                    s.assign(
-                        BranchName.of("this-one-should-not-exist"),
-                        Optional.empty(),
-                        s.noAncestorHash())),
-        new ReferenceNotFoundFunction("assign/hash")
-            .msg("Commit '12341234123412341234123412341234123412341234' not found")
-            .function(
-                s ->
-                    s.assign(
-                        BranchName.of("main"),
-                        Optional.empty(),
-                        Hash.of("12341234123412341234123412341234123412341234"))),
-        // delete()
-        new ReferenceNotFoundFunction("delete/branch")
-            .msg("Named reference 'this-one-should-not-exist' not found")
-            .function(s -> s.delete(BranchName.of("this-one-should-not-exist"), Optional.empty())),
-        new ReferenceNotFoundFunction("delete/tag")
-            .msg("Named reference 'this-one-should-not-exist' not found")
-            .function(s -> s.delete(BranchName.of("this-one-should-not-exist"), Optional.empty())),
-        // create()
-        new ReferenceNotFoundFunction("create/hash")
-            .msg("Commit '12341234123412341234123412341234123412341234' not found")
-            .function(
-                s ->
-                    s.create(
-                        BranchName.of("foo"),
-                        Optional.of(Hash.of("12341234123412341234123412341234123412341234")))),
-        // commit()
-        new ReferenceNotFoundFunction("commit/branch")
-            .msg("Named reference 'this-one-should-not-exist' not found")
-            .function(
-                s ->
-                    s.commit(
-                        BranchName.of("this-one-should-not-exist"),
-                        Optional.empty(),
-                        "meta",
-                        singletonList(Delete.of(Key.of("meep"))))),
-        new ReferenceNotFoundFunction("commit/hash")
-            .msg(
-                "Could not find commit '12341234123412341234123412341234123412341234' in reference 'main'.")
-            .function(
-                s ->
-                    s.commit(
-                        BranchName.of("main"),
-                        Optional.of(Hash.of("12341234123412341234123412341234123412341234")),
-                        "meta",
-                        singletonList(Delete.of(Key.of("meep"))))),
-        // transplant()
-        new ReferenceNotFoundFunction("transplant/branch/ok")
-            .msg("Named reference 'this-one-should-not-exist' not found")
-            .function(
-                s ->
-                    s.transplant(
-                        BranchName.of("this-one-should-not-exist"),
-                        Optional.empty(),
-                        singletonList(s.hashOnReference(BranchName.of("main"), Optional.empty())),
-                        Function.identity())),
-        new ReferenceNotFoundFunction("transplant/hash/empty")
-            .msg(
-                "Could not find commit '12341234123412341234123412341234123412341234' in reference 'main'.")
-            .function(
-                s ->
-                    s.transplant(
-                        BranchName.of("main"),
-                        Optional.of(Hash.of("12341234123412341234123412341234123412341234")),
-                        singletonList(Hash.of("12341234123412341234123412341234123412341234")),
-                        Function.identity())),
-        new ReferenceNotFoundFunction("transplant/empty/hash")
-            .msg("Commit '12341234123412341234123412341234123412341234' not found")
-            .function(
-                s ->
-                    s.transplant(
-                        BranchName.of("main"),
-                        Optional.empty(),
-                        singletonList(Hash.of("12341234123412341234123412341234123412341234")),
-                        Function.identity())),
-        // merge()
-        new ReferenceNotFoundFunction("merge/hash/empty")
-            .msg("Commit '12341234123412341234123412341234123412341234' not found")
-            .function(
-                s ->
-                    s.merge(
-                        Hash.of("12341234123412341234123412341234123412341234"),
-                        BranchName.of("main"),
-                        Optional.empty(),
-                        Function.identity())),
-        new ReferenceNotFoundFunction("merge/empty/hash")
-            .msg(
-                "Could not find commit '12341234123412341234123412341234123412341234' in reference 'main'.")
-            .function(
-                s ->
-                    s.merge(
-                        s.noAncestorHash(),
-                        BranchName.of("main"),
-                        Optional.of(Hash.of("12341234123412341234123412341234123412341234")),
-                        Function.identity())));
+  @Nested
+  protected class Assign extends AbstractAssign {
+    public Assign() {
+      super(AbstractVersionStoreTestBase.this.store());
+    }
   }
 
-  @ParameterizedTest
-  @MethodSource("referenceNotFoundFunctions")
-  void referenceNotFound(ReferenceNotFoundFunction f) throws Exception {
-    if (f.setup != null) {
-      f.setup.run(store());
+  @Nested
+  protected class Transplant extends AbstractTransplant {
+    public Transplant() {
+      super(AbstractVersionStoreTestBase.this.store());
     }
-    assertThatThrownBy(() -> f.function.run(store()))
-        .isInstanceOf(ReferenceNotFoundException.class)
-        .hasMessage(f.msg);
   }
 
-  /** Assigning a branch/tag to a fresh main without any commits didn't work in 0.9.2 */
-  @Test
-  public void assignReferenceToFreshMain()
-      throws ReferenceNotFoundException, ReferenceAlreadyExistsException,
-          ReferenceConflictException {
-    BranchName main = BranchName.of("main");
-    WithHash<Ref> mainRef = store().toRef(main.getName());
-    assertThat(store().getCommits(main, false)).isEmpty();
-    try (Stream<ReferenceInfo<String>> refs = store().getNamedRefs(GetNamedRefsParams.DEFAULT)) {
-      assertThat(refs).extracting(r -> r.getNamedRef().getName()).containsExactly(main.getName());
+  @Nested
+  protected class Merge extends AbstractMerge {
+    public Merge() {
+      super(AbstractVersionStoreTestBase.this.store());
     }
+  }
 
-    BranchName testBranch = BranchName.of("testBranch");
-    Hash testBranchHash = store().create(testBranch, Optional.empty());
-    store().assign(testBranch, Optional.of(testBranchHash), mainRef.getHash());
-    assertThat(store().toRef(testBranch.getName()).getHash()).isEqualTo(mainRef.getHash());
+  @Nested
+  protected class Diff extends AbstractDiff {
+    public Diff() {
+      super(AbstractVersionStoreTestBase.this.store());
+    }
+  }
 
-    TagName testTag = TagName.of("testTag");
-    Hash testTagHash = store().create(testTag, Optional.empty());
-    store().assign(testTag, Optional.of(testTagHash), mainRef.getHash());
-    assertThat(store().toRef(testTag.getName()).getHash()).isEqualTo(mainRef.getHash());
+  @Nested
+  public class ReferenceNotFound extends AbstractReferenceNotFound {
+    public ReferenceNotFound() {
+      super(AbstractVersionStoreTestBase.this.store());
+    }
   }
 }
